+++ frontend/src/lib/components/canisters/InstallWAppDetails.svelte
